requirements.txt
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
joblib==1.3.2
scipy==1.11.1
jupyter==1.0.0

# MLOps Stack
apache-airflow==2.8.0
apache-airflow-providers-docker==3.7.0
mlflow==2.8.0
fastapi==0.104.0
uvicorn==0.24.0
streamlit==1.28.0
prometheus-client==0.19.0
docker==6.1.0

# ML Models
xgboost==2.0.0

# Data Versioning
dvc==3.30.0

# Additional utilities
pyyaml==6.0.1
requests==2.31.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.7 
prometheus-client